// admin/src/lib/countries.generated.ts
export type CountryInfo = {
  name: string;
  lat: number | null;
  lon: number | null;
};